libview: checking if point is inside annotation
Added a step to check if a point is inside an
annotation. In case it is foud to be inside the
bounding box, view asks the document if it is
actually inside the annotations, since the
quadrilaterals might not comprise the whole are of
the bounding box. Only then it become selectable.

https://bugzilla.gnome.org/show_bug.cgi?id=769133

Fixes #696